+++ src/main/java/com/nix/repository/ReportRepository.java
@@ -7,9 +7,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.nix.models.Report;
 
 public interface ReportRepository extends JpaRepository<Report, Long> {
-	public List<Report> findByCommentId(Integer commentId);
+	public List<Report> findByCommentId(Long commentId);
 
-	public List<Report> findByChapterId(Integer chapterId);
+	public List<Report> findByChapterId(Long chapterId);
 
-	public List<Report> findByBookId(Integer bookId);
+	public List<Report> findByBookId(Long bookId);
 }
