+++ src/main/java/com/nix/service/CommentService.java
@@ -12,33 +12,33 @@ public interface CommentService {
 
 	public List<Comment> getAllParentComments();
 
-	public Page<Comment> getPagerBookComments(int page, int size, Integer bookId);
+	public Page<Comment> getPagerBookComments(int page, int size, Long bookId);
 
-	public Page<Comment> getPagerChapterComments(int page, int size, Integer chapterId);
+	public Page<Comment> getPagerChapterComments(int page, int size, Long chapterId);
 
-	public Page<Comment> getPagerPostComments(int page, int size, Integer postId);
+	public Page<Comment> getPagerPostComments(int page, int size, Long postId);
 	
-	public Page<Comment> getRecentCommentsByUserId(Integer userId, int page, int size);
+	public Page<Comment> getRecentCommentsByUserId(Long userId, int page, int size);
 
-	public Comment createPostComment(Comment comment, Integer postId, User user) throws Exception;
+	public Comment createPostComment(Comment comment, Long postId, User user) throws Exception;
 
-	public Comment createBookComment(Comment comment, Integer bookId, User user) throws Exception;
+	public Comment createBookComment(Comment comment, Long bookId, User user) throws Exception;
 
-	public Comment createChapterComment(Comment comment, Integer chapterId, User user) throws Exception;
+	public Comment createChapterComment(Comment comment, Long chapterId, User user) throws Exception;
 
-	public Boolean likeComment(Integer commentId, Integer userId) throws Exception;
+	public Boolean likeComment(Long commentId, Long userId) throws Exception;
 	
-	public Boolean isCommentLikedByCurrentUser(Integer commentId, User user);
+	public Boolean isCommentLikedByCurrentUser(Long commentId, User user);
 
-	public Comment findCommentById(Integer commentId) throws Exception;
+	public Comment findCommentById(Long commentId) throws Exception;
 
-	public String deleteComment(Integer commentId, Integer userId) throws Exception;
+	public String deleteComment(Long commentId, Long userId) throws Exception;
 
-	public Comment editComment(Integer userId, Integer commentId, Comment comment) throws Exception;
+	public Comment editComment(Long userId, Long commentId, Comment comment) throws Exception;
 
-	public Comment createReplyBookComment(Comment comment, Integer parentCommentId, User user) throws Exception;
+	public Comment createReplyBookComment(Comment comment, Long parentCommentId, User user) throws Exception;
 
-	public Comment createReplyChapterComment(Comment comment, Integer parentCommentId, User user) throws Exception;
+	public Comment createReplyChapterComment(Comment comment, Long parentCommentId, User user) throws Exception;
 
-	public Comment createReplyPostComment(Comment comment, Integer parentCommentId, User user) throws Exception;
+	public Comment createReplyPostComment(Comment comment, Long parentCommentId, User user) throws Exception;
 }
